+++ source/blender/render/intern/source/initrender.c
@@ -157,7 +157,7 @@ float RE_filter_value(int type, float x)
 		{
 			const float two_gaussfac2 = 2.0f * gaussfac * gaussfac;
 			x *= 3.0f * gaussfac;
-			return 1.0f / sqrtf(M_PI * two_gaussfac2) * expf(-x*x / two_gaussfac2);
+			return 1.0f / sqrtf((float)M_PI * two_gaussfac2) * expf(-x*x / two_gaussfac2);
 		}
 			
 		case R_FILTER_MITCH:
